requirements.txt
pyyaml
locust
requests